requirements.txt
asgiref==3.2.3
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
pycparser==2.19
PyPDF2==1.26.0
Pyphen==0.9.5
python3-openid==3.1.0
pytz==2019.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
tinycss2==1.0.2
urllib3==1.25.8
WeasyPrint==51
webencodings==0.5.1
xhtml2pdf==0.2.4
